+++ demo/web/components/accordion/main.dart
@@ -1,9 +1,8 @@
 import 'package:angular/angular.dart';
-import 'accordion_demo.dart';
 
 // ignore: uri_has_not_been_generated
-import 'main.template.dart' as ng_generated;
+import 'accordion_demo.template.dart' show AccordionDemoNgFactory;
 
 main() {
-  bootstrapStatic(AccordionDemo, [], ng_generated.initReflector);
+  runApp(AccordionDemoNgFactory);
 }