+++ demo/web/components/input/main.dart
@@ -1,9 +1,8 @@
 import 'package:angular/angular.dart';
-import 'input_demo.dart';
 
 // ignore: uri_has_not_been_generated
-import 'main.template.dart' as ng_generated;
+import 'input_demo.template.dart' show InputDemoNgFactory;
 
 main() {
-  bootstrapStatic(InputDemo, [], ng_generated.initReflector);
+  runApp(InputDemoNgFactory);
 }