+++ demo/web/components/modal/main.dart
@@ -1,9 +1,8 @@
 import 'package:angular/angular.dart';
-import 'modal_demo.dart';
 
 // ignore: uri_has_not_been_generated
-import 'main.template.dart' show initReflector;
+import 'modal_demo.template.dart' show ModalDemoNgFactory;
 
 main() {
-  bootstrapStatic(ModalDemo, [], initReflector);
+  runApp(ModalDemoNgFactory);
 }