+++ demo/web/components/prompt/main.dart
@@ -1,9 +1,8 @@
 import 'package:angular/angular.dart';
-import 'prompt_demo.dart';
 
 // ignore: uri_has_not_been_generated
-import 'main.template.dart' show initReflector;
+import 'prompt_demo.template.dart' show PromptDemoNgFactory;
 
 main() {
-  bootstrapStatic(PromptDemo, [], initReflector);
+  runApp(PromptDemoNgFactory);
 }