+++ demo/web/components/table/main.dart
@@ -5,11 +5,11 @@ import 'table_demo.dart';
 import 'package:dson/dson.dart';
 
 // ignore: uri_has_not_been_generated
-import 'main.template.dart' as ng_generated;
+import 'table_demo.template.dart' show TableDemoComponentNgFactory;
 
 part 'main.g.dart';
 
 main() {
   _initMirrors();
-  bootstrapStatic(TableDemoComponent, [], ng_generated.initReflector);
+  runApp(TableDemoComponentNgFactory);
 }