+++ demo/web/components/tabs/main.dart
@@ -1,9 +1,8 @@
 import 'package:angular/angular.dart';
-import 'tabs_demo.dart';
 
 // ignore: uri_has_not_been_generated
-import 'main.template.dart' show initReflector;
+import 'tabs_demo.template.dart' show TabsDemoNgFactory;
 
 main() {
-  bootstrapStatic(TabsDemo, [], initReflector);
+  runApp(TabsDemoNgFactory);
 }