+++ demo/web/components/tabsx/main.dart
@@ -1,9 +1,8 @@
 import 'package:angular/angular.dart';
-import 'tabsx_demo.dart';
 
 // ignore: uri_has_not_been_generated
-import 'main.template.dart' show initReflector;
+import 'tabsx_demo.template.dart' show TabsxDemoNgFactory;
 
 main() {
-  bootstrapStatic(TabsxDemo, [], initReflector);
+  runApp(TabsxDemoNgFactory);
 }