+++ demo/web/components/timepicker/main.dart
@@ -1,9 +1,8 @@
 import 'package:angular/angular.dart';
-import 'timepicker_demo.dart';
 
 // ignore: uri_has_not_been_generated
-import 'main.template.dart' show initReflector;
+import 'timepicker_demo.template.dart' show TimepickerDemoNgFactory;
 
 main() {
-  bootstrapStatic(TimepickerDemo, [], initReflector);
+  runApp(TimepickerDemoNgFactory);
 }